+++ benches/multi_thread/delivery.rs
@@ -176,9 +176,9 @@ fn spsc( c: &mut Criterion )
 			(
 				move || // setup
 				{
-					let (sum_in_addr, sum_in_mb) = Addr::builder().bounded( Some(BOUNDED) ).build() ;
+					let (sum_in_addr, sum_in_mb) = Addr::builder( "sum_in" ).bounded( Some(BOUNDED) ).build() ;
+					let (sum_addr, sum_mb)       = Addr::builder( "sum"    ).bounded( Some(BOUNDED) ).build() ;
 					let sum                      = Sum{ total: 5, inner: sum_in_addr }              ;
-					let (sum_addr, sum_mb)       = Addr::builder().bounded( Some(BOUNDED) ).build() ;
 
 					let sumin_thread = thread::spawn( move ||
 					{
@@ -207,7 +207,7 @@ fn spsc( c: &mut Criterion )
 
 						let res = sum_addr.call( Show{} ).await.expect( "Call failed" );
 
-						assert_eq!( *msgs as u64 *10 + 5 + termial( *msgs as u64 ), res );
+						assert_eq!( *msgs *10 + 5 + termial( *msgs ), res );
 					});
 
 					sumin_thread.join().expect( "join sum_in thread" );
@@ -266,7 +266,7 @@ fn spsc( c: &mut Criterion )
 
 		// 				let res = sum_addr.call( Show{} ).await.expect( "Call failed" );
 
-		// 				assert_eq!( *msgs as u64 *10 + 5 + termial( *msgs as u64 ), res );
+		// 				assert_eq!( *msgs *10 + 5 + termial( *msgs ), res );
 		// 			});
 
 		// 			sumin_thread.join().expect( "join sum_in thread" );
@@ -304,7 +304,7 @@ fn spsc( c: &mut Criterion )
 
 						let res = sum_addr.send( Show{} ).await.expect( "Call failed" );
 
-						assert_eq!( *msgs as u64 *10 + 5 + termial( *msgs as u64 ), res );
+						assert_eq!( *msgs *10 + 5 + termial( *msgs ), res );
 
 						sum_in_thread.stop();
 						sum_thread   .stop();
@@ -327,9 +327,9 @@ fn spsc( c: &mut Criterion )
 			(
 				move || // setup
 				{
-					let (sum_in_addr, sum_in_mb) = Addr::builder().bounded( Some(BOUNDED) ).build() ;
-					let sum      = Sum{ total: 5, inner: sum_in_addr }                              ;
-					let (sum_addr, sum_mb) = Addr::builder().bounded( Some(BOUNDED) ).build()       ;
+					let (sum_in_addr, sum_in_mb) = Addr::builder( "sum_in" ).bounded( Some(BOUNDED) ).build() ;
+					let (sum_addr   , sum_mb   ) = Addr::builder( "sum"    ).bounded( Some(BOUNDED) ).build() ;
+					let sum                      = Sum{ total: 5, inner: sum_in_addr }                        ;
 
 
 					let sumin_thread = thread::spawn( move ||
@@ -359,7 +359,7 @@ fn spsc( c: &mut Criterion )
 
 						let res = sum_addr.call( Show{} ).await.expect( "Call failed" );
 
-						assert_eq!( *msgs as u64 *10 + 5 + termial( *msgs as u64 ), res );
+						assert_eq!( *msgs *10 + 5 + termial( *msgs ), res );
 					});
 
 					sumin_thread.join().expect( "join sum_in thread" );
@@ -419,7 +419,7 @@ fn spsc( c: &mut Criterion )
 
 		// 				let res = sum_addr.call( Show{} ).await.expect( "Call failed" );
 
-		// 				assert_eq!( *msgs as u64 *10 + 5 + termial( *msgs as u64 ), res );
+		// 				assert_eq!( *msgs *10 + 5 + termial( *msgs ), res );
 		// 			});
 
 		// 			sumin_thread.join().expect( "join sum_in thread" );