+++ examples/concurrent.rs
@@ -90,7 +90,7 @@ impl Handler<Ping> for MyActor
 async fn main() -> Result< (), DynError >
 {
 	let actor    = MyActor{ exec: Box::new(AsyncStd) };
-	let mut addr = Addr::builder().spawn( actor, &AsyncStd )?;
+	let mut addr = Addr::builder( "concurrent" ).spawn( actor, &AsyncStd )?;
 
 	// Admittedly, this looks a bit weird. Call is fallible, and it returns a result over
 	// the SpawnError, since the handler needs to spawn and spawning is fallible.