+++ CHANGELOG.md
@@ -5,9 +5,29 @@
 [Unreleased]: https://github.com/najamelan/async_executors/compare/release...dev
 
 
+## [0.3.0] - 2022-07-01
+
+[0.3.0]: https://github.com/najamelan/async_executors/compare/0.2.1...0.3.0
+
+### Upgraded
+
+  - __BREAKING__: Update thespis to 0.2.0.
+  - __BREAKING__: Change the `ActorBuilder::channel` signature so you don't
+                  have to manually convert the error type on the `Sink`.
+
+### Changed
+
+  - __BREAKING__: Names have become mandatory on `Mailbox` and `ActorBuilder`.
+  - __BREAKING__: Improved ergonomics of using custom channels.
+  - switch to edition 2021
+
+### Added
+  - more examples and tests
+
+
 ## [0.2.1] - 2022-05-26
 
-[0.2.0]: https://github.com/najamelan/async_executors/compare/0.2.0...0.2.1
+[0.2.1]: https://github.com/najamelan/async_executors/compare/0.2.0...0.2.1
 
 ### Fixed